modetest: util: Add support for P01[026] and S01[026] formats
Adds support for buffer allocation in the P010, P012, and P016
2-plane YUV420 10, 12, and 16 bit formats, and the S010, S012, and
S016 3-plane YUV420 formats of the same bit depths.

As all the pattern generation uses 8bit values, store a shift
value for each format to denote how the 8 bit value needs to
be altered to store in a 16bit word.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>