util: pattern: Fix chroma offset for insert_value_yuv_planar
The chroma offset calculation resulted in the running beyond
the end of the buffer.

eg For ysub=2, y=0 and y=1 should result in writing the same U & V
line, but chroma_offset = (y + 1) / ysub; resulted in them going
to different lines, potentially running off the end of the buffer
and causing a seg fault.

Update the calculation so lines 0 to (ysub-1) all write the same
output line.

Fixes: 40aeab6fd5b8 ("modetest: util: pattern: add new patterns")
Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>